Add Conduit-typed message definitions
Define message queue traffic with concrete message types, implemented as
Conduit structs.

These structs define the messages as they currently exist (with some
simplifications when possible). The message and topic structure around
Relay communication can stand to be simplified more (but that requires
changes in Relay, so it can wait). Additionally, the structure of
users/requestors and rooms in messages can be further constrained (right
now, they're just maps), but that can happen as the adapter interface
itself is improved.

Some other highlights:
* Removed old `Cog.Command.Request` and `Cog.Command.Response` structs
* Removed last vestiges of dynamic configuration code from Cog This old
  code was from the earliest implementation of dynamic configuration,
  and predates Relay, I believe.
* Removed some dead helper code
* Remove `id` field from adapter response messages
  Turns out this wasn't actually being used except in tests. We can
  actually remove it and tests can still pass.