Bumped embedded bundle version
Internal testing has shown the embedded bundle version in some folks'
databases was upgraded by the code which tried to detect embedded bundle
version changes (since removed). This situation shows up as a failure to
start up Cog as it refuses to downgrade the embedded bundle.

Since we're not 100% sure if user installations could be in a similar
state we've decided to play it safe and bump the version number one more time.