Sync cogctl and cog in the Docker image
Previously, if you didn't edit the `scripts/docker-build` file
beforehand, you could build an image that had the "wrong" version of
`cogctl` included. This version would always be build from `master` of
`cogctl`, and not necessarily the version that corresponded with the
`cog` you were building.

Now, the script figures out what branch `cog` is checked out to, and if
`cogctl` has a branch of the same name, it will check `cogctl` out on
that branch. If it doesn't, then it will fall back to the `master`
branch in `cogctl`.

Previously, we had `.git` in our `.dockerignore` file, but that would
make it impossible to determine `cog`'s current branch. Now, we don't
ignore `.git`, but we do remove it from the image before we finish
building. We also do the same for `cogctl` (which we weren't before).