State explicitly how many responses to assert against
Because of how the tests and helpers were originally laid out, there
would be intermittent failures for tests where we expect two responses
back, as happens when we send back a message saying a user has been
autoregistered, in addition to the output of the pipeline they
initiated. Depending on the timing of the tests and network activity
between us and Slack, we might get back just the confirmation, or the
confirmation and the actual response.

Now we explicitly declare how many messages we expect to assert against,
and continue to poll until we have that many messages.