Accumulate multiple values of list options
Command options that are list-typed can now be declared multiple times,
with all values being collected into a single list.

That is,

    do:something --foo=one --foo=two --foo=three

will end up with `["one", "two", "three"]` as the value for the `foo`
option.

The previous behavior of splitting comma-delimited strings to obtain the
list values is still supported. Thus `do:something --foo=one,two,three`
is equivalent to the above command.

The two behaviors can be combined, as well:

    do:something --foo=one --foo=two,three,four --foo=five

results in `["one", "two", "three", "four", "five"]` being the value for
`foo`.