Don't crash if a table has no rows
TableRex (for whatever reason) treats a table with no rows as an
error. We feel differently. In the case that a table doesn't have any
rows, we'll just manually generate an empty table.

Additionally, Greenbar won't actually recognize a table without
headers (it's just text), so we can tighten up the contract for
rendering a table; the first child will always be a `table_header`. If
we somehow manage to get a table that doesn't conform to that,
processing will fall through to our catch-all clause.