Aborts no longer crash the executor
Previously, if a command aborted the pipeline, we would end up
crashing in the executor. When we introduced the new templates, we
added a guard to the `respond` function. When we processed an aborted
command, we weren't capturing the information needed to satisfy the
guard.

We also no longer add the pipeline ID to the `AuditMessage` call,
because it already includes that information in the general audit
message body.