Trigger cogctl pipeline
This is now a pipeline step, rather than a post-command callback, in
order to ensure that it fires only after all preceeding steps pass.

For now, we're not requiring the cogctl pipeline to pass in order for
this cog pipeline to pass, but that can easily be added with a
subsequent API polling step in the future.