Optionally add chat handle for bootstrap user
When setting up a Cog system for the first time using `COG_BOOTSTRAP_*`
variables, a `COG_BOOTSTRAP_CHAT_HANDLE` variable may optionally be
set. The bootstrapping process will attempt to resolve this handle
according to the currently-active chat provider (controlled by
`COG_SLACK_ENABLED` and `COG_HIPCHAT_ENABLED`).

If the given handle is valid, it is associated with the bootstrapped
user. That user can immediately begin interacting with Cog via chat.

If the chat handle is not valid, a warning message is logged, but the
bootstrapped user is still created and the system still starts up as
usual. In this case, the user can use `cogctl` to set up a chat handle.