Configure public URLs via env vars
This introduces the following environment variables as a way to
configure public-facing URLs generated by Cog:

* COG_API_URL_BASE
* COG_TRIGGER_URL_BASE
* COG_SERVICE_URL_BASE

Each of these variables may be a full URL of the form
`scheme://host:port/path`, which will be used as a prefix for all
generated URLs handled by the corresponding Phoenix endpoint. This is
provided instead of the alternative, which would expose each of the four
URL components (per endpoint!) as a separate variable.

This will enable users to run Cog behind a proxy by allowing them to
generate URLs relative to the proxy, and not the particular server that
Cog is hosted on.

These environment variables are optional; if they are missing, URLs will
be generated according to the configuration of the existing endpoint
listeners.

Previously, we provided `*_URL_HOST` and `*_URL_PORT` environment
variables for this purpose. We did not provide variables for scheme and
path, though, so this was an incomplete solution for users wanting to do
SSL termination, or present all of Cog behind a unified URL scheme on a
single port. These variables, if present, will still be honored, but
they are now deprecated, and will be removed in a future release of
Cog. If they are used, a deprecation message is logged, but it does not
use the same formatting as other Cog log messages, since the processing
of the variables must take place at compile-time and not runtime, when
our Logger configuration is in effect. This is (sadly) necessary,
because Phoenix will not allow the `path` component of a URL
configuration to be altered at runtime (`scheme`, `host`, and `port`,
however, may be altered at any time).

The default Logger format is in effect for these deprecation messages:

```
21:53:41.165 [warn]  The environment variable 'COG_API_URL_PORT' is deprecated and will cease to be honored in an upcoming release. Please transition to using 'COG_API_URL_BASE' as soon as you can.
```

Our Docker Compose files have been updated to use the new variables.

Fixes #1158