increase slack API wait interval to avoid throttling
Per the Slack documentation:
> In general we allow applications that integrate with Slack to send no more than one message per second. We allow bursts over that limit for short periods. However, if your app continues to exceed the limit over a longer period of time it will be rate limited.

With our previous interval of 750ms we'd see inconsistent results. Sometimes we'd be able to burst enough to finish tests without being limited, and sometimes we wouldn't. This should err on the side of being conservative and allow tests to complete at the expense of being a little bit slower. :(