Revise Slack tests to account for API throttling
Slack will throttle an account if it makes more than ~1 request per
second. Since the Slack library we're using currently doesn't employ any
backpressure to help deal with this, we're relying on sleeping to deal
with this.

The throttling kept happening with the Slack user account we are
using, and not the bot account. In addition to changing how we sleep,
we now use the built-in state of the Slack client to get user
information, which eliminates extra API calls.

Additionally, we're now using the same `ProviderCase` for all our Slack
tests. This should hopefully make it easier to refine these tests in the
future.

Some cleanup / refactoring of the `SlackClient` code was also done.