Revise DataStoreTest setup
Since Mix / ExUnit start up the Cog application, and the `DataStore`
process is a registered one, we need to "swap out" that process for
tests.

This should be done for each test, swapping back the original process at
the end of each one.

Tests are also no longer run asynchronously, to prevent odd race
conditions.