Increase event timestamp precision and accuracy
Previously, the timestamps we generated for Cog's event log had only
second resolution, despite the fact that events frequently occur at the
microsecond scale. This made it more complicated that it should be to
order events (e.g., when analyzing data using a log aggregator).

Now, we log using ISO8601 timestamps with microsecond resolution.

In the course of making this change, lots of time-related code was
removed, in favor of native Elixir code (which didn't exist when the
code was initially written). We still need to have custom code to
compute `DateTime` differences, though. Sigh.

It was also discovered that our elapsed microsecond event log data
didn't quite line up with the timestamps we were printing; this is
because we weren't being consistent with what we were using as our basis
of comparison. Now, all elapsed time computations are done within the
actual event creation code, instead of externally, and are all based on
a consistent "start time". This simplifies the executor a small bit. We
also no longer emit an info-level logging statement stating how long a
pipeline took to execute, but that information is available from the
event log (and *that* log is structured data, rather than plain text).

We get not only increased precision, but increased accuracy as well!
BONUS!

Fixes #1253